Split runtime option parser policy
