+++ substratevm/src/com.oracle.svm.interpreter/src/com/oracle/svm/interpreter/ristretto/RistrettoUtils.java
@@ -46,7 +46,7 @@ import com.oracle.svm.core.graal.meta.SubstrateReplacements;
 import com.oracle.svm.core.meta.SharedType;
 import com.oracle.svm.core.hub.DynamicHub;
 import com.oracle.svm.core.log.Log;
-import com.oracle.svm.core.option.RuntimeOptionParser;
+import com.oracle.svm.core.option.RuntimeOptionParserPolicy;
 import com.oracle.svm.guest.staging.option.RuntimeOptionValues;
 import com.oracle.svm.graal.RuntimeCompilationSupport;
 import com.oracle.svm.graal.SubstrateGraalUtils;
@@ -462,7 +462,7 @@ public class RistrettoUtils {
                 // Use name=value boolean format for compatibility with Graal options
                 CommonOptionParser.BooleanOptionFormat booleanFormat = CommonOptionParser.BooleanOptionFormat.NAME_VALUE;
                 for (String option : options) {
-                    RuntimeOptionParser.singleton().parseOptionAtRuntime(option, "", booleanFormat, values, false);
+                    RuntimeOptionParserPolicy.parseOptionAtRuntime(option, "", booleanFormat, values, false);
                 }
             }
